# apps/hermes-studio/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hermes Studio API",
    "description": "Hermes Studio API — chat sessions, scheduled jobs, platform channels, model management, skills, memory, logs, file browser, group chat, and terminal.",
    "version": "0.6.30"
  },
  "servers": [
    {
      "url": "http://localhost:8648",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/workbench/paper-recommendations": {
      "get": {
        "tags": [
          "Workbench"
        ],
        "summary": "Get paper-recommendations",
        "description": "GET /api/workbench/paper-recommendations",
        "operationId": "getPaper-recommendations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "tags": [
          "Workbench"
        ],
        "summary": "Create paper-recommendations",
        "description": "POST /api/workbench/paper-recommendations",
        "operationId": "createPaper-recommendations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/workbench/paper-recommendations/refresh": {
      "post": {
        "tags": [
          "Workbench"
        ],
        "summary": "Create refresh",
        "description": "POST /api/workbench/paper-recommendations/refresh",
        "operationId": "createRefresh",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/workbench/summary": {
      "get": {
        "tags": [
          "Workbench"
        ],
        "summary": "Get summary",
        "description": "GET /api/workbench/summary",
        "operationId": "getSummary",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/workbench/wechat-discovery": {
      "post": {
        "tags": [
          "Workbench"
        ],
        "summary": "Create wechat-discovery",
        "description": "POST /api/workbench/wechat-discovery",
        "operationId": "createWechat-discovery",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/workbench/wechat-import": {
      "post": {
        "tags": [
          "Workbench"
        ],
        "summary": "Create wechat-import",
        "description": "POST /api/workbench/wechat-import",
        "operationId": "createWechat-import",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/workbench/wechat-sources": {
      "get": {
        "tags": [
          "Workbench"
        ],
        "summary": "Get wechat-sources",
        "description": "GET /api/workbench/wechat-sources",
        "operationId": "getWechat-sources",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "tags": [
          "Workbench"
        ],
        "summary": "Create wechat-sources",
        "description": "POST /api/workbench/wechat-sources",
        "operationId": "createWechat-sources",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/workbench/wechat-sources/sync": {
      "post": {
        "tags": [
          "Workbench"
        ],
        "summary": "Create sync",
        "description": "POST /api/workbench/wechat-sources/sync",
        "operationId": "createSync",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/workbench/wechat-sources/{id}": {
      "delete": {
        "tags": [
          "Workbench"
        ],
        "summary": "Delete :id",
        "description": "DELETE /api/workbench/wechat-sources/:id",
        "operationId": "deleteWechat-sources",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad Request - Invalid parameters",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request"
                }
              }
            }
          }
        }
      },
...
    }
  }
}